Restore index exports
