+++ src/index.ts
@@ -1,22 +1,4 @@
 export * from "./builder";
-export {
-	AlignXCenter,
-	AlignXCenterTo,
-	AlignXLeft,
-	AlignXRight,
-	AlignYBottom,
-	AlignYCenter,
-	AlignYTop,
-	BackgroundOp,
-	DistributeX,
-	DistributeY,
-	type LayoutOperator,
-	type NodeRecord,
-	type StackAlignment,
-	type StackChild,
-	StackH as StackHOp,
-	StackV as StackVOp,
-} from "./operators";
 export * from "./output";
 export * from "./parser";
 export * from "./solver";
