+++ src/solver/align.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
+import { solveLayout } from ".";
 import type { LayoutOperator, NodeRecord } from "./operators";
 import { AlignXLeft } from "./operators";
-import { solveLayout } from "./solver";
 
 describe("Align operator", () => {
 	it("aligns multiple nodes to the leftmost X", () => {
