+++ src/solver/operators.ts
